requirements.txt
absl-py==0.7.1
alabaster==0.7.12
asn1crypto==0.24.0
astor==0.8.0
Babel==2.7.0
certifi==2019.6.16
cffi==1.12.3
chardet==3.0.4
Click==7.0
cloudpickle==1.2.1
cryptography==2.7
cycler==0.10.0
cytoolz==0.10.0
dask==2.3.0
decorator==4.4.0
docutils==0.15.2
easydict==1.9
Flask==1.1.1
gast==0.2.2
grpcio==1.22.0
h5py==2.9.0
idna==2.8
imageio==2.5.0
imagesize==1.1.0
imgaug==0.2.9
itsdangerous==1.1.0
Jinja2==2.10.1
joblib==0.13.2
Keras==2.2.4
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
Markdown==3.1.1
MarkupSafe==1.1.1
matplotlib==3.1.1
mkl-fft==1.0.14
mkl-random==1.0.2
mkl-service==2.0.2
networkx==2.3
numpy==1.16.4
olefile==0.46
packaging==19.0
Pillow==6.1.0
protobuf==3.8.0
pycparser==2.19
Pygments==2.4.2
pyOpenSSL==19.0.0
pyparsing==2.4.2
pyreadline==2.1
PySocks==1.7.0
python-dateutil==2.8.0
pytz==2019.2
PyWavelets==1.0.3
PyYAML==5.1.2
requests==2.22.0
requests-toolbelt==0.9.1
scikit-image==0.15.0
scikit-learn==0.21.2
scipy==1.3.1
Shapely==1.6.4.post2
six==1.12.0
snowballstemmer==1.9.0
Sphinx==2.2.0
sphinxcontrib-applehelp==1.0.1
sphinxcontrib-devhelp==1.0.1
sphinxcontrib-htmlhelp==1.0.2
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.2
sphinxcontrib-serializinghtml==1.1.1
tensorboard==1.14.0
tensorflow==1.14.0
tensorflow-estimator==1.14.0
termcolor==1.1.0
toolz==0.10.0
tornado==6.0.3
tqdm==4.35.0
urllib3==1.24.2
waitress==1.2.1
Werkzeug==0.15.5
win-inet-pton==1.1.0
wincertstore==0.2
wrapt==1.11.2
